html and css framework for recipes
This is just a rough skeleton of the recipes page. The colour scheme is not what was voted for but just random ones, we will do it later when all the info is on the pages. Please dont touch the code for the menu at the side, it took me a while to get it to work. Replace all the pictures and text, all the code for the structure is set up so just need to insert the urls and text between the codes.
If you know how to get rid of the massive space at the top of the page, please do (the title is too lower). Hyperlinks doesn't work yet but I'll set it up later.  Design is the same for all of them, if you have suggestions on how to make it better, please do do it but just make sure to tell us so we can apply it to the other recipe pages. Don't forget to push, commit and comment after every working change so the others can follow as well. Good Luck!