background and hyperlinks on front page
I removed the underlines on the hyperlinks and coded it so that when you hover over it it changes colour and when you click it, it changes to another colour.
Background is of my dog until we decide on one. I made it so it won't repeat, covers the page and is centred.